Incident regression in r (#1)
This is the first version of my incident uptake regression method for
uptake projections.
- R contains the functions
- scripts contains a short script using the R functions
- data contains .csv files with full names and abbreviations of
geographic regions (state and country)

The primary aim is to import data into a standardized format, fit a
regression model on incident uptake, and use that regression model to
make projections (either genuine predictions of the future or
predictions of the data the model was trained on). There are also
functions for interpolating dates and scaling one data set to another.

@Fuhan-Yang please review to determine if this works for now (given the
impending switch to Python), or if you'd like any changes. @swo please
also feel free to review and/or comment. Thank you!