Created a mostly-empty incident_regression python package (#2)
I used poetry to create the scaffold for a (currently empty) python
package, with just a few dependencies that I know of so far - polars,
datetime, and numpy. This is where we want to focus our effort on
building well-structured code for uptake projections based on incident
uptake regression models.