First Python version with simple architecture (#4)
This is the Python version of my original R, consisting of three
functions (in /iup) and a script that calls them sequentially (in
/scripts). Either IIS or NIS data is imported and formatted, a linear
regression model is fit to the per-day incident uptake, and projections
from this linear model are generated. Projections are returned as a
DataFrame in the same format as the original data.

This code is meant to be minimal. Many features (some of which were
present in the R version) are omitted here, e.g.:
- scaling between data sets
- interpolation of dates
- ability to specify any model other than: uptake ~ 1 + time_elapsed +
previous_uptake + time_elapsed*previous_uptake
- group-level variation on state or calendar year
- quantification of projection uncertainty
- any graphical anything

The purpose of this PR is to get a minimal Python version in our shared
space. That way, @Fuhan-Yang has a starting point for building
evaluation metrics in Python, while I can build in more features and
smarter architectures with the help of @swo.