Add data description (#24)
Hi Scott and Ed, 

This is an atomic PR to add the data description in the README.md and
delete the `NIS_2023-24.csv`, which can be acquired through API.

This is the public API that returns the same data as in the csv:
"https://data.cdc.gov/resource/pakc-hru3.json". The `get_uptake_data()`
currently does not have the `.json` input option, maybe this can be
added in the future development.