Next Attempt at Object-Oriented Architecture (#34)
Comments on my first attempt at an object-oriented architecture (#6)
warranted a total overhaul. Here is my new attempt. Many of the comments
from #6 are addressed, and the rest I recorded as issues (not linked to
PR#6, though, because I closed that). Plenty of improvement is still
necessary, but I think I'm ready for the next round of feedback.