Generalization for arbitrary grouping factors (#53)
This uses the .over() approach to accommodate arbitrary grouping
factors. The user now specifies in the config .yaml file which grouping
factors to use from each data set, giving both the names of the relevant
NIS columns as well as the desired names for these columns. The desired
names for the grouping factor columns ought to match across any data
sets that will be pooled together into the same model. A new function
`extract_group_names` verifies this.

This PR does NOT include any testing - I will save that for later. 

So, please ignore PRs #49, #50, and #51. These should be kept a little
while longer for reference, but not merged. This PR is the one that
should be thoroughly reviewed and merged to permit arbitrary grouping
factors.