Refactor data type and forecast type (#64)
![module_component](https://github.com/user-attachments/assets/d9dcf9c1-13b0-4174-8675-37476f3af6be)
 
This PR deletes the intermediate classes between ValidatedData and the
end data types for observed data (Incident/CumulativeUptakeData) and
forecasts (Quantile/SampleForecast). Each class has slightly different
validation methods. In this way, the ValidatedData is a pool of
functions, where different functions are used by multiple subclasses.
The pro of this architecture is it is clear and easy to understand,
while the con is that some functions are not meaningful for some classes
but they can still be called. This is explicit for `split_train_test`
function, which should not be available for any forecast types. Maybe
consider to put it under both Incident&CumulativeUptakeData. However,
other functions like 'date_to_*' can be useful for forecast types as
well.

Another note is that, the outcome from `predict()` in the
LinearIncidentUptakeModel should be PointForecast type to feed into
metric function.

The UptakeData is replaced with ValidatedData in `test_uptake_data.py`,
please correct if anything is done wrong!