Refactored Linear Incident Uptake Model to return projections separately, rather than as attributes (#72)
Note that the LinearIncidentUptakeModel must return cumulative
projections. The reason is that if projections begin mid-season,
cumulative projections will include the uptake that is already known,
but incident projections won't. So cumulative projections are returned,
and from there they can be converted to incident (as shown in the
main.py script).

I also fixed a few issues in the current main branch that Pylance
flagged on my local machine.

Resolves #31