R script to explore models (#90)
I have written an R script to explore a few model structures with full
posterior predictions. This R script is not intended to be a
user-friendly final product; it is merely personal scratchwork.
Moreover, there are some scientific shortcomings - most notably some
foolishly naive priors for both models described below. But its
preliminary exploration shows a few results worth noting.

The data span the first historical date through 2024-11-30, for both
covid and flu. Two model structures are explored:
- Linear Incident Model (LIM): incident uptake this week is linearly
related to incident uptake last week and to time-since-rollout. This
model predicts entire trajectories of uptake.
- Cumulative Hill Model (CHM): cumulative uptake follows the shape of a
Hill function. This model predicts single data points of uptake.

I have chosen a forecast date of 2024-10-25. All data before this are
used to train the model. All data after this (2024-10-25 to 2024-11-30,
approx. 1 month) are used for evaluation. Both models predict through
2025-05-31, well into the future. Thus, we obtain quantitative
evaluation for a ~1-month horizon, and visualization for a much longer
horizon. You can choose different forecast dates and horizons in lines
25-26 of the script.

Here are all the available data for covid:

![image](https://github.com/user-attachments/assets/bb54e47e-a91e-40ac-9a82-8e679085c5bd)

And here are the predictions from each model in the 2024/2025 season for
covid:

![image](https://github.com/user-attachments/assets/71c09e1c-ef67-448c-941a-a9c2febfbb1a)
Mean squared percent error on the ~1-month horizon: LIM 1.15, CHM 6.03
Absolute error on 2024-11-30: LIM -0.48%, CHM 2.34%

Here are all the available data for flu:

![image](https://github.com/user-attachments/assets/c8879197-1496-466c-9d7e-3f7de190aafe)

And here are the predictions from each model in the 2024/2025 season for
flu:

![image](https://github.com/user-attachments/assets/77548cfa-f9e0-4f44-be8a-30ef0345f1ce)
Mean squared percent error on the ~1-month horizon: LIM 2.26, CHM 23.52
Absolute error on 2024-11-30: LIM -0.17%, CHM -3.50%

At least one issue is immediately visible for each model:
- LIM: for rather long prediction horizons (~6+ months), predictions can
be unstable (see flu predictions 300+ days post-rollout). This is
because the model predicts trajectories, in which each week's incident
uptake correlates with the next's. If a trajectory's incident uptake
gets too high or low, "positive feedback" of sorts can take over,
leading to runaway predictions in the high or low direction. This might
be addressed with smarter priors.
- CHM: predictions are disjoint from the most recent data (see both flu
and covid predictions on the forecast date (dashed vertical line)). This
is because the model only captures the _average_ cumulative uptake
trajectory across all seasons, without understanding that entire seasons
differ from one another systematically. This might be addressed with
group-level effects of season, but trying this breaks MCMC currently, so
it will also require smarter priors. Or we might hack this by separately
rescaling each season's cumulative uptake on a 0-1 scale. A related
problem that these approaches will _not_ solve is that uncertainty is
constant in time, rather than appropriately increasing into the future.
(Note that the LIM shown here also ignores group-level effects of
season, but because it models incident rather than cumulative uptake, it
naturally picks up where the training data leave off.)

This script should facilitate exploration of adjustments and entirely
new models, and it will help us sanity check the same models in numpyro.