Pipeline setup step 2: Fill forecast.py (#93)
This PR adds forecast function and run the forecast sequentially given
the time frame in config.yaml. The results are saved as parquet data in
the designated path. While this function hasn't been tested because of
several errors:

1. Error: "typeFloat64 is incompatible with expected type String" is
returned in `split_train_test`. This happens when the uptake data is a
single polars DataFrame (instead of a uptake data list).

2. The NaN in training data prevents the fitting of
`LinearIncidenceUptakeModel`, checks or methods (impute NA, or delete)
are needed to ensure fitting can be run.

---------

Co-authored-by: Scott Olesen <ulp7@cdc.gov>