Hyperparameters for season in the Hill model (in R demo only) (#117)
I have updated the R demo to incorporate hyperparameters into the Hill
model that allow the final uptake and the half-maximal date to vary by
season. Because this is just in the R demo, which is for brainstorming
and will not be part of the final codebase, it doesn't require a
detailed review. This is just to demonstrate few points, using the flu
data.

As a reminder, here is what the entirety of the available flu data looks
like:

![image](https://github.com/user-attachments/assets/7e72d57a-8b86-488b-901c-015c901c27ed)

If a Hill model is fit to these data, and then a forecast for the
2024/25 season is made from late October 2024 onward, we get:

![image](https://github.com/user-attachments/assets/70fc645c-c718-4265-83d4-e5f65b58fc74)
Notice how 1) the forecast does not align with the data on the forecast
date, and 2) the 95% credible interval starts off wide without widening
much into the future.

Now we do the same thing, but including hyperparameters to allow the
exact shape of the Hill curve to vary by season:

![image](https://github.com/user-attachments/assets/e75094f7-f199-4b66-804f-73964dc3e33b)
Both problems are nearly solved! The forecast aligns closely (though not
perfectly) with the data on the forecast date, and the credible interval
starts off much thinner (though still encompassing some values lower
than observed cumulative uptake on the forecast date).

We can probably ignore the last vestiges of these two problems, because
we are moving toward a refactor in which we no longer trust the observed
data to be synonymous with the true latent uptake. In that case, some
uncertainty about what the true uptake is, even on dates that have
already been observed, is totally acceptable.

I think this is a nice validation of @swo's intuition for how prediction
uncertainty should be greater at the tail of a sigmoid curve as opposed
to the middle, and a nice validation of our collective intuition that
grouping the data on season is important.