Add individual projections and interval in plotting (#140)
- Add `plot_individual_projections` in `postprocess.py` to allow
plotting individual samples vs data.
- Add `plot_summary` in `postprocess.py` to allow plotting mean, 95%
credible interval with data.
Currently, line plot is used instead of area because the challenge of
using alt.area on pivotted quantile data. This is my trial code (no
error but only returned blank):

```
interval = alt.Chart(plot_new_data).transform_filter(
       datum.type == 'pred'
).transform_quantile(
      "estimate",probs = [0.0025,0.5,0.975],as_=["quantile","estimate"],  # Calculate quantile for each grouping factor #
       groupby=["time_end","season","forecast_start","model"]
).transform_pivot(
      "quantile", groupby=["time_end"], value="estimate" # Convert long data frame to wide data frame #
).mark_area(opacity=0.3,color = "blue").encode(
        x="time_end:T",
        y="0.025:Q",
        y2="0.975:Q",
)

```
I think the error is that the new columns is named as a quantile number
(0.0025,0.975), since a plot was returned if the new column name is a
string. Similar question has been asked
[online](https://stackoverflow.com/questions/68515371/plot-area-between-two-quantiles-in-altair)
but hasn't been resolved.

Another way is that we manually transform the data frame and feed to
altair. But before that I would like to know if you have any ideas to
fix this!

- Modify Makefile to allow generating plot directly. The commented line
will be useful and added once the interval metrics are added.

If running smoothly, two plots are expected:

Individual plot (400 trajectories):


![image](https://github.com/user-attachments/assets/db71e0c7-d630-411a-84f5-14aa3562fa19)

Summary plot: 


![image](https://github.com/user-attachments/assets/cdf33ab0-e557-4365-a19a-d4dbcce9069d)