Refactor Hill Model (#141)
The Hill model is refactored to handle arbitrary grouping factors
(including but not limited to season), separate observation from latent
true uptake, etc. See docs/model_details.md for the full description of
this model.

This is currently a work in progress. In particular, I am running into
`Segmentation fault (core dumped)` when I run the model on flu data
across 14 seasons and 54 states/jurisdictions. I did not expect this to
be a large-enough data set or a complex-enough model to crash my CPU, so
more investigation must follow.