More careful standardization (#149)
The standardization utility now handles NaNs consistently and does not
apply the Bessel correction, regardless of whether the input is a numpy
ndarray or a polars expression. I tried to streamline the function to
look cleaner, but lurking around every corner was some operation or
conditional that required separating the ndarray vs. expression cases
anyway. I don't think this is worth any more effort.

This will address #128