Refactor Hill Model, again (#147)
The Hill model is improved in two important ways:
- The observation-layer likelihood is now a beta-binomial, rather than a
truncated Normal
- The true latent uptake is modeled by a Hill function _plus_ a linear
function, rather than a Hill alone

This necessitates a change in preprocessing, by which each empirical
point estimate and 95% confidence interval is translated into the total
number of people surveyed and the (smaller) number of whom are
vaccinated. To do this, we assume that the reported 95% CIs are Wald
intervals on the standard error of the mean (SEM) for the point
estimate. As such, the SEM is the half-width of the reported 95% CI
divided by ~1.96. Letting $p = \text{point estimate of uptake}$, we
solve the following for N:

```math
SEM = \sqrt{\frac{(1-p) \cdot p^2 + p \cdot (1-p)^2}{N}}
```

This allows us to calculate the total number of people surveyed,
$N_{tot}$, and the number of people vaccinated, $N_{vax}$, for each data
point.

The refactored Hill + Linear model fits the data much more easily: the
MCMC chains move and produce posterior samples that are reasonable at
first glance, even when grouping factors like season and state are
included. There is likely more tinkering yet to be done (e.g. ESS is low
for some parameters), but this will probably be easier dealt with once
MCMC diagnostics and faceted forecast plots by grouping factor are
built.

Please feel free to ignore `docs/model_journal.md` and
`scripts/model_demo.R`! These are for my own record-keeping and sloppy
scratchwork, respectively.