Posterior predictive check plots (#153)
This requires no code review - I've merely replaced the outdated stuff
in my scratchwork script ("model_demo.r") with some rudimentary
posterior predictive plots. There is plenty to improve, but I think this
shows we are on the right track.

I load the flu data and the logistic+linear model posterior that this
repo produces as a .parquet (see #143). For each season from 2009/2010
through 2021/2022, I plot the mean predicted latent true uptake vs. the
data across states. (The 95% credible interval on the latent true uptake
is also included, but it is usually quite thin.)

Note that the model was fit with 1 chain (not enough) using 500 warmups
(not enough) and 500 samples (not enough), so there is room for
immediate improvement.

Here's an example of the 2016-2017 season.


![image](https://github.com/user-attachments/assets/db581e73-3d0b-47e6-9de7-42fff488bd1e)

Since this requires no code review, we can just merge it, but I wanted
everyone to see where we are currently.

cc @Fuhan-Yang @afmagee42 @swo