Parallelized numpyro chains (#167)
If there are multiple MCMC chains for model fitting (and there should
be), they now run in parallel. MCMC control parameters are now separated
from model parameters in the config to set up the parallelization a
little more easily.

This resolves issue #120. 

This also resolves issue #160, because that Arviz warning was basically
saying that we should run more than 1 MCMC chain.

This also resolves issue #110, because I also split the random seed that
initializes the model into two children: one for model fitting and one
for prediction.