Adjusted Priors in Small but Crucial Ways (#173)
I adjusted the suggested priors in the config template to give better
MCMC behavior. The crucial change, I believe, is increasing the gamma
prior for M (slope of the linear function) so that the prior probability
approaches 0 as values of M approach 0. That is, while M is likely
small, we are confident that it is also greater than 0.

Before making this change, 16 independent HMC chains across 4 different
random seeds ranged in run times from 14 to 78 minutes, and the longer
chains produced hundreds of divergences on rare occasions.

After this change, 20 independent HMC chains across 5 different random
seeds ranged in run times from 10 to 13 minutes and produced at most 2
(but usually 0) divergences.

This is a tiny amount of code, but a big improvement. This addresses
#171