Plot trajectories, quantile summary, and evaluation with Streamlit (#183)
- Updated config to use NIS flu data
- Add a script to launch streamlit app

To run streamlit, add streamlit in poetry, and `make viz` in terminal.
Refreshing is needed occasionally to load the plot.

This streamlit app allows user to select the dimensions they want to
see, and other dimensions will be used to filter the data. For example:
the users can observe the color-labeled forecast trajectories in Alabama
made on different forecast dates by column, on different seasons by
rows.

There are three tabs, each showing: individual trajectories, 95% PI with
mean estimate and data, evaluation score (mspe, absolute difference)
over forecast start date. If it's not loading, refresh the web browser.

Need to update: 
1. Currently, if users choose geography in one of the dimensions, all
the observed data across states will show in each faceted plot, instead
of the data only from that state. This is more an Altair problem. When
there is a layered and faceted plot (multiple charts of data and
prediction), it is supposed to only show the value under the
corresponding group factors for both data frames. One solution is that
to aggregate the data with the forecasts, and only deal with a single
data frame. But that will lose the feature that observed data is showed
as points, while forecasts are showed as lines, which can be important
distinction in the individual trajectory plot.
2. Currently, the date from metric, the absolute difference on a
specific date is selected from config. Maybe we can let the user to
choose in streamlit? But that can be a more later problem.
3. More like a future upgrade: when mouse hovers to the points of
observed data, it shows the detailed value. How to make this possible
for forecast too? This interactive feature will make the user's
experience better!

At least, the default dimensions look great! We finally have an app to
show our results!

Resolves #174