Posterior Checks (#182)
While making plots for a demo presentation, I realized:
1) We generate forecasts to compare vs. the test data, but no posterior
checks to compare vs. the training data.
2) The model's forecasts are substantially worse than the model's fit to
training data.
3) The pipeline at times feels clunky, and it can be tricky to not
overwrite results that took a long time to get.

This PR only deals with 1. I will make issues for 2 (#185) and 3 (#186).

Essentially, I added posterior checks as an output of the `forecast.py`
script, and I adjusted the Makefile accordingly.

The other altered file, `figs_demo.py`, is my notebook-in-progress to
make custom figures for the demo presentation that sparked all this. No
need to review this in great detail, since we may end up pruning it out
of the repo later. But if you want to take a look and comment on whether
we should make lasting use of this (e.g. as a demo that lives in the
repo), feel free.