Fixed Bug in Forecasting (#194)
We previously noticed that although the model performed well on training
data, it performed terribly when forecasting. I have traced this problem
back to a bug in the model fitting/prediction code and fixed it. The
changes to `iup/models.py` and `iup/utils.py` are the key changes.

Feel free to ignore the mass of sloppy code in `scripts/figs_demo.py` -
this is merely the scratchwork I used to make plots as I investigated
the aforementioned bug. I want to let this scratchwork hang around a few
weeks, in case any of it proves repeatedly useful. But most of it will
eventually disappear, so it is not worth your time to review.

I will circulate a more detailed explanation of my model diagnostic
process in another channel, which will give the context for why I've
made the changes I did here. Also happy to discuss it synchronously.