Nicknames in the Makefile (#195)
Line 1 of the Makefile is now a nickname that is assigned to a
particular model fitting + forecasting + scoring run. After you `make
all`, if you `make all` again without changing the nickname, Make only
runs steps for which a target and/or prerequisite has changed, as
expected. But if you `make all` again with a different nickname, the
entire pipeline runs again, and its output is stored without overwriting
the previous run. This addresses #186.

To accomplish this, when `make all` is run with a new nickname, 5 new
subdirectories (named using the nickname) are created within the 5
output directories (data, fits, diagnostics, forecasts, and scores) to
hold the corresponding outputs. Just a little extra work using `Path`
from `pathlib` in each of the scripts enables the dynamic creation of
these directories. Placeholder files are not required, and many are
removed, so this addresses #189.

`make clean` will remove all 5 of these nicknamed subdirectories.

There are also several other minor changes for the sake of convenience:
- The `plot` vs. `table` distinction in the output directories is
removed, reducing the number of targets that need to be defined in the
Makefile and the number of arguments that several of the scripts need.
- The interface with the NIS API in the Makefile is simplified, so that
`make nis` will cache all available NIS API data sets (warning and
skipping any that already exist in the cache, so as to not overwrite),
and `make delete_nis` will clear the cache of all NIS API data sets.
- The cache for the NIS API data sets is now the global system default
(typically `/home/<userid>/.cache/nisapi`) to avoid the need for a local
NIS data cache inside the repo.
- The Makefile's default config file `config_template.yaml` is changed
to only perform one forecast, not multiple on different forecast dates.
This is merely for faster testing of the pipeline. Multiple forecasts
for evaluation purposes can be added back in at any time.
- Many of the packages have been updated, including updating the NIS API
to the latest version, as reflected in `poetry.lock`.
- Some `CumulativeUptakeData` and `QuantileForecast` objects must be
coerced into standard data frames in the `summarize_score` function of
`iup/eval.py` to make their types match for joining. This wasn't a
problem before, so I suspect that updating the package version of polars
introduced this new requirement.

This is more for @Fuhan-Yang to review than @afmagee42, but I thought
you both might be interested in the nicknaming strategy.