Scratchwork plots for presentation (#199)
NO REVIEW REQUIRED

In the script `figs_demo.py` I've whipped up some figures to use for
presentation purposes. This is scratchwork and does not require a
review. We probably just want to merge this to have it stored for
temporary reference, until the streamlit visualizations are complete and
the presentation is over.

I've also deleted the old `model_demo.r` script, which was also
scratchwork and is no longer useful.