Clarify config (#204)
@eschrom Not sure if you had a different vision, but these were the
changes I had to make in order for this repo to make sense to me.

Point Makefile to `scripts/config.yaml`, which is gitignore'd. In the
README, tell users to copy `scripts/config_template.yaml` to this
location and manipulate it. Add a note in `eval.py` that makes clear why
evaluation isn't done, if it's not being done.