More GAM exploration (#209)
More exploratory analysis of GAM. 

Mathematically, the GAM model we used is: 
$Y = X\beta+ Zu + \beta_0$

$Y$ is the observed vaccine uptake, $X$ is the design matrix of basis
function with N * k dimension, where N is the number of observations and
k is the number of basis functions used. Each element in $X$ is the
value of the basis function evaluated at the predictor `elapsed`.
$\beta$ is a vector of coefficients that control each basis function.
$X\beta$ is the main effect that is the same across all the level in a
group (in our case is all seasons in `season` ).

$Z$ is the covariance matrix of each grouping factor, and $u$ is
group-specific intercept. $Z$ is an identity matrix that assumes each
level has effect on $Y$ independent from other levels. $Zu$ is the
random effect that models the level-specific intercept.

$beta_0$ is the main intercept. 

When fitting the model with national flu vaccine uptake, we found that
the fitted line by each season shares the same shape and the intercept
differs by season.

![fitted_value_by_season](https://github.com/user-attachments/assets/99b134fe-9e85-421f-ac57-0b0bc8cab41f)

To confirm that, if we exclude the group-specific intercept, the main
effect fitted by each season is the same

![main_effect_by_season](https://github.com/user-attachments/assets/3cc74bba-e0c0-4c77-95b5-952790e5fd8c)

When predicting, the predicted group-specific intercept is 0 for the
unobserved group, resulting that only the main effect trained by
observed seasons is showed:

![pred_by_sel_seasons](https://github.com/user-attachments/assets/bdd4eadc-06af-49dd-ae0d-b7d5d5ce65ad)
Here we respectively used 2, 4, 8, 12, and 14 seasons in training and
showed the prediction.

It is expected to see that MSPE decreased as more seasons are included
in the training data.

![mspe_by_sel_seasons](https://github.com/user-attachments/assets/433c7d0a-74b6-44d5-aec7-a7ed262f90ed)

We also tried GAM on Covid-19 vaccine uptake. The training data showed
goodness-of fit for the observed uptake:

![covid_train](https://github.com/user-attachments/assets/d8528802-ce04-4cb3-9fc0-f1e19126b1f2)

And the prediction showed reasonable forecast (starting from Sep 1,
2024), with MSPE 0.0012.

![covid_test](https://github.com/user-attachments/assets/2ed6dbd9-9d9c-498f-8bfe-51331f44d6c9)