EDA on number of basis functions given cumulative and incidental uptake data (#211)
Before switching to Bayesian framework, I did a quick EDA based on
today's discussion: how does gam perform given different number of basis
functions on cumulative and incidental data?

Conditions: 
1. Flu vaccine uptake data from NIS is used.  
2. Model is assumed to be log-normal distributed (log link function)
3. Cubic basis spline function is used
4. Incidence data is calculated as the difference of cumulative data,
and the first "0" point is trimmed to ensure all the data is eligible
for log link function.
5. Training and test data is split on 2023-07-01. Season 2023/2024 is
used for test.

Methods:
1. GAM model is fitted given the number of basis functions (k) varied
between 5 and 15.
2. Fitted values and predictions are plotted for each k.
3. MSPE is calculated between test data and prediction for each k. 

Results: 

![gam_k_cumu_fitted](https://github.com/user-attachments/assets/4d3b685e-d44a-4f21-a454-d2a95d1350d8)
As we found out in the morning that more basis functions are needed to
better fit the smooth shape of the cumulative data, this also applies to
prediction:

![gam_k_cumu_pred](https://github.com/user-attachments/assets/424d7796-3b3e-4d92-8382-6cd9c5b1ccf9)
k needs to be at least 10 to fit and predict the smooth sigmoid shape of
the cumulative data.

Thus, followed by the thought that incidental data may need fewer basis
functions as it includes the turning shape by itself, we've used the
same method on incidental data:

![gam_k_inc_fitted](https://github.com/user-attachments/assets/1a21b6a9-3afd-4fd0-9bf3-cac8adf28537)
Interestingly, it is found that k is needed to be at least 8 to fit the
peak of the incidental vaccine uptake. I speculate that more basis
function still provides higher flexibility to model the sharp turning
(peak) of the incidence data.

This also applies to prediction: 

![gam_k_inc_pred](https://github.com/user-attachments/assets/452bf277-53a5-4799-91bd-454301102bbd)

But the peak can be fitted and predicted when k is at least 8, which is
lower than the k for cumulative data. Thus, our assumption is valid that
it is easier to capture the wiggliness of the incidence data than
cumulative data.

Also confirmed this using MSPE between test and prediction across k for
incidence and cumulative data:

![all_mspe_gam_k](https://github.com/user-attachments/assets/f58b093c-08d2-4c13-867e-96037199d4de)

We can see that incidence data has pronounced advantage in predicting
vaccine uptake. Given the results, I will move forward with trying
incidence data on the Bayesian framework, first.

I also searched if there is basis function that is similar to sigmoid
function, but because the basis functions are naturally polynomials,
there is no basis function that has sigmoid shape. Increasing k to
achieve the sigmoid shape of the smooth function is the common approach,
as we expected. This also supports us to move forward with incidence
data.