GAM draft in numpyro (#213)
- Added a document to describe the math of GAM inference in Bayesian
framework.
- Added a script to implement the math in the document.

So far, only the main effect (uptake ~s(elapsed)) is considered. Random
effect will be added once the inference on main effect is reviewed and
improved.

Currently, the inference logic is that, given fixed knots (14) and
degree(3,i.e, cubic spline), there are 10 basis functions. Here, a
package `skfda` (based on scipy) is used to derive the design matrix and
the penalty matrix (integrated second-order derivative). Design matrix,
penalty matrix and vaccine uptake(estimate) are used to infer the 10
coeffcients ($\beta$)of the 10 basis functions respectively. Other
parameters to infer are:
- $\lambda$, the smoothing parameter that controls the level of penalty
- $\sigma$, one of parameters to control the covariance matrix in the
prior of $\beta$

The numpyro failed when running, because that the inverse of penalty
matrix could be negative, invalidating the covariance matrix of the
multivariate normal distribution of $\beta$. The priors of $\lambda$ and
$\sigma$ are chosen without much knowledge and subject to change.

There can be other issues that I fail to see...this is a pretty drafted
version, any comments/critique/ this is not going to work give up/ are
welcome and will be well-accepted :)

My main references:
1.
https://www.annualreviews.org/content/journals/10.1146/annurev-statistics-112723-034249
2.
https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.14498