Clarify dependencies (#225)
- Move dev dependencies into the dev dependency group
- Keep GAM dependency as an optional group
- Specify an NIS API version floor

See [PEP 735](https://peps.python.org/pep-0735/) which explains that
*optional* dependencies are for optional functionality in the codebase
(which is how a GAM would be) while dependency *groups* are for things
that aren't part of the package (e.g., pytest is used for developing the
package but is never called while the package is running per se).

[This thread](https://github.com/astral-sh/uv/issues/9011) also has a
good explanation.